+++ capture_gui/plugins/viewportplugin.py
@@ -1,5 +1,4 @@
 from capture_gui.vendor.Qt import QtCore, QtWidgets
-
 import capture_gui.plugin
 import capture_gui.lib as lib
 import capture
@@ -17,33 +16,64 @@ class ViewportPlugin(capture_gui.plugin.Plugin):
     def __init__(self, parent=None):
         super(ViewportPlugin, self).__init__(parent=parent)
 
-        self.show_type_actions = list()
-        self.show_types = lib.get_show_object_types()
-
+        # set inherited attributes
         self.setObjectName(self.label)
 
+        # custom atttributes
+        self.show_type_actions = list()
+
+        # get information
+        self.show_types = lib.get_show_object_types()
+
+        # set main layout for widget
         self._layout = QtWidgets.QVBoxLayout()
         self._layout.setContentsMargins(0, 0, 0, 0)
         self.setLayout(self._layout)
+
+        # build
+        # region Menus
+        menus_vlayout = QtWidgets.QHBoxLayout()
+
+        # Display Lights
+        self.display_light_menu = self._build_light_menu()
+        self.display_light_menu.setFixedHeight(20)
+
+        # Show
+        self.show_types_button = QtWidgets.QPushButton("Show")
+        self.show_types_button.setFixedHeight(20)
+        self.show_types_menu = self._build_show_menu()
+        self.show_types_button.setMenu(self.show_types_menu)
+
+        # fill layout
+        menus_vlayout.addWidget(self.display_light_menu)
+        menus_vlayout.addWidget(self.show_types_button)
+
+        # endregion Menus
+
+        # region Checkboxes
+        checkbox_layout = QtWidgets.QGridLayout()
+        self.high_quality = QtWidgets.QCheckBox()
+        self.high_quality.setText("Force Viewport 2.0 + AA")
         self.override_viewport = QtWidgets.QCheckBox("Override viewport "
                                                      "settings")
         self.override_viewport.setChecked(True)
 
-        # region Show
-        self.show_types_button = QtWidgets.QPushButton("Show")
-        self.show_types_button.setFixedWidth(150)
-        self.show_types_menu = self._build_show_menu()
-        self.show_types_button.setMenu(self.show_types_menu)
-        # endregion Show
+        # two sided lighting
+        self.two_sided_ligthing = QtWidgets.QCheckBox("Two Sided Ligthing")
+        self.two_sided_ligthing.setChecked(False)
 
-        # region Checkboxes
-        self.high_quality = QtWidgets.QCheckBox()
-        self.high_quality.setText("Force Viewport 2.0 + AA")
+        # show
+        self.shadows = QtWidgets.QCheckBox("Shadows")
+        self.shadows.setChecked(False)
+
+        checkbox_layout.addWidget(self.override_viewport, 0, 0)
+        checkbox_layout.addWidget(self.high_quality, 0, 1)
+        checkbox_layout.addWidget(self.two_sided_ligthing, 1, 0)
+        checkbox_layout.addWidget(self.shadows, 1, 1)
         # endregion Checkboxes
 
-        self._layout.addWidget(self.override_viewport)
-        self._layout.addWidget(self.show_types_button)
-        self._layout.addWidget(self.high_quality)
+        self._layout.addLayout(checkbox_layout)
+        self._layout.addLayout(menus_vlayout)
 
         # signals
         self.high_quality.stateChanged.connect(self.options_changed)
@@ -61,7 +91,7 @@ class ViewportPlugin(capture_gui.plugin.Plugin):
         menu = QtWidgets.QMenu(self)
         menu.setObjectName("ShowShapesMenu")
         menu.setWindowTitle("Show")
-        menu.setFixedWidth(150)
+        menu.setFixedWidth(180)
         menu.setTearOffEnabled(True)
 
         # Show all check
@@ -84,11 +114,36 @@ class ViewportPlugin(capture_gui.plugin.Plugin):
 
         return menu
 
+    def _build_light_menu(self):
+        """
+        Create the menu items for the different types of lighting for
+        in the viewport
+        
+        :return: None 
+        """
+
+        menu = QtWidgets.QComboBox(self)
+
+        # names cane be found in
+        display_lights = {"Use Default Lighting": "default",
+                          "Use Flat Lighting": "flat",
+                          "Use Selected Lights": "active",
+                          "Use All Lights": "all",
+                          "Use No Lights": "none"}
+
+        for label, name in display_lights.items():
+            menu.addItem(label, userData=name)
+
+        return menu
+
     def on_toggle_override(self):
         """Enable or disable show menu when override is checked"""
         state = self.override_viewport.isChecked()
         self.show_types_button.setEnabled(state)
         self.high_quality.setEnabled(state)
+        self.display_light_menu.setEnabled(state)
+        self.shadows.setEnabled(state)
+        self.two_sided_ligthing.setEnabled(state)
 
     def toggle_all_visbile(self):
         """Set all object types off or on depending on the state"""
@@ -119,6 +174,18 @@ class ViewportPlugin(capture_gui.plugin.Plugin):
 
         return show_inputs
 
+    def get_displaylights(self):
+        """
+        Get the currently selected displayLight
+        
+        :return: the system name of the selected displayLight
+        :rtype: dict
+        """
+        indx = self.display_light_menu.currentIndex()
+        return {"displayLights": self.display_light_menu.itemData(indx),
+                "shadows": self.shadows.isChecked(),
+                "twoSidedLighting": self.two_sided_ligthing.isChecked()}
+
     def get_inputs(self, as_preset):
         """
         Return the widget options
@@ -131,7 +198,10 @@ class ViewportPlugin(capture_gui.plugin.Plugin):
         :rtype: dict
         """
         inputs = {"high_quality": self.high_quality.isChecked(),
-                  "override_viewport_options": self.override_viewport.isChecked()}
+                  "override_viewport_options": self.override_viewport.isChecked(),
+                  "displayLights": self.display_light_menu.currentIndex(),
+                  "shadows": self.shadows.isChecked(),
+                  "twoSidedLighting": self.two_sided_ligthing.isChecked()}
 
         inputs.update(self.get_show_inputs())
 
@@ -150,14 +220,21 @@ class ViewportPlugin(capture_gui.plugin.Plugin):
         # get input values directly from input given
         override_viewport = inputs.get("override_viewport_options", True)
         high_quality = inputs.get("high_quality", True)
+        displaylight = inputs.get("displayLights", 0)
+        two_sided_ligthing = inputs.get("twoSidedLighting", False)
+        shadows = inputs.get("shadows", False)
 
         self.high_quality.setChecked(high_quality)
         self.override_viewport.setChecked(override_viewport)
         self.show_types_button.setEnabled(override_viewport)
 
+        # display light menu
+        self.display_light_menu.setCurrentIndex(displaylight)
+        self.shadows.setChecked(shadows)
+        self.two_sided_ligthing.setChecked(two_sided_ligthing)
+
         for action in self.show_type_actions:
-            label = action.text()
-            system_name = self.show_types[label]
+            system_name = self.show_types[action.text()]
             state = inputs.get(system_name, True)
             action.setChecked(state)
 
@@ -185,9 +262,12 @@ class ViewportPlugin(capture_gui.plugin.Plugin):
                 outputs['viewport2_options']['multiSampleCount'] = 8
 
             show_per_type = self.get_show_inputs()
+            display_lights = self.get_displaylights()
             outputs['viewport_options'].update(show_per_type)
+            outputs['viewport_options'].update(display_lights)
         else:
             # Use settings from the active viewport
+            # NOTE: WHen this fails we should give the user a warning
             outputs = capture.parse_active_view()
 
         # TODO: we could filter out the settings we want to use or leave it be